Migrated all managed lists into a component.
- Component loader now imports all files
- Clear previously used __init__ imports
