--- flexget/components/backlog/__init__.py
@@ -1,3 +1,0 @@
-# Import plugins when component is imported
-from . import cli
-from . import backlog
