--- flexget/components/history/__init__.py
@@ -1,4 +1,0 @@
-# Import plugins when component is imported
-from . import api
-from . import cli
-from . import history
