--- flexget/components/irc/__init__.py
@@ -1,3 +1,0 @@
-# Import plugins when component is imported
-from . import cli
-from . import irc
