+++ flexget/components/managed_lists/list_clear.py
