+++ flexget/components/managed_lists/list_remove.py
