--- flexget/components/seen/__init__.py
@@ -1,6 +1,0 @@
-# Import plugins when component is imported
-from . import api
-from . import cli
-from . import seen
-from . import seen_movies
-from . import seen_info_hash
