--- flexget/components/trakt/__init__.py
@@ -1,7 +1,0 @@
-# Import plugins when component is imported
-from . import api
-from . import cli
-from . import api_trakt
-from . import next_trakt_episodes
-from . import trakt_calendar
-from . import trakt_lookup
