+++ flexget/tests/test_pending_list.py
@@ -1,8 +1,8 @@
 from __future__ import unicode_literals, division, absolute_import
 from builtins import *  # pylint: disable=unused-import, redefined-builtin
 
+from flexget.components.managed_lists.lists.pending_list.db import PendingListList
 from flexget.manager import Session
-from flexget.plugins.list.pending_list import PendingListList
 
 
 class TestListInterface(object):
